Moved services & guards into separate folders
