+++ src/app/log-in/log-in.component.ts
@@ -1,4 +1,4 @@
-import { LogInService } from './../services/log-in.service';
+import { LogInService } from '../services/log-in-service/log-in.service';
 
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Component, OnInit } from '@angular/core';
