+++ src/app/services/log-in-service/log-in.service.ts
@@ -1,4 +1,4 @@
-import { ILogIn } from './../log-in/log-in.interface';
+import { ILogIn } from '../../log-in/log-in.interface';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { throwError, of } from 'rxjs';
